web_wol_client/requirements.txt
flask>=2.3.0
flask-sqlalchemy>=3.0.0
flask-login>=0.6.0
flask-wtf>=1.1.0
wtforms>=3.0.0
werkzeug>=2.3.0
bcrypt>=4.0.0
wakeonlan>=3.0.0
ping3>=4.0.0
python-nmap>=0.7.0
schedule>=1.2.0
gunicorn>=21.0.0
